vncc-bot/src/requirements.txt
pyTelegramBotAPI
python-dotenv
requests
Pillow
